Bugfix: type coersion for CP leveling check.
Using vconfig to turn off cp leveling by setting cpexp to 0 causes type errors in the
noexp comparison because it turns the table value from an int to a string. Add explicit
tonumber() to force cpexp to a number during comparison. (This error broke the enemy HP
gauge as a side effect.)